Talents reworked from the ground up
Previously, all Talents could be casted in any direction; for both technical and design reasons, Talents can now only be cast left, right, or in whatever special direction the Talent specifies. The consequences are as follows:

TargettingController is now redundant, and has been removed.
The Marauder will no longer attempt to jump up in order to attack its target.

Additional changes were also made as part of this rework:

Talent & emission statuses have been crunched down to a 3-state integer.
Talent.Initialize() now only has the owner param, and gets components from that.
Talents have a TelegraphEffect() along with CastEffect(), for use in Talents that do something on telegraph start (such as Slash). This eliminates the need to override Cast() in children of Talent.
An animation for a Slash with a sabre-type weapon was added, and the Slash AOE sprite was changed accordingly.
Casting a Talent stops movement for its duration.
FIX: Marauder previously did not override its parent Awake().